Added Tiling script

+++ torrentvideos.sh
@@ -6,14 +6,13 @@ read -p "Connect to nordvpn? (y/n): " -n 1 -r
 echo ""
 if [[ $REPLY =~ ^[Yy]$ ]]
 then
-    nordvpn c > /dev/null
+    nordvpn c #> /dev/null
 fi
 
 pirate-get --save-magnets $search
 
 magnetfile=$(ls)
 magnetlink=$(cat "$magnetfile")
-rm "$magnetfile"
 
 if [ "$magnetfile" == "" ]
 then
@@ -21,6 +20,7 @@ then
     exit 1
 fi
 
+rm "$magnetfile"
 # transmission-cli end when downloaded
 tmpfile=$(mktemp)
 chmod a+x $tmpfile
@@ -32,23 +32,29 @@ transmission-cli --download-dir . -f $tmpfile $magnetlink 2> /dev/null
 
 # Move from torrentdir to parentdir
 torrentfolder=$(ls)
-cd "$torrentfolder"
-mv * ../
-cd ..
-rm -r "$torrentfolder"
+if [ -d "$torrentfolder" ]
+then
+    cd "$torrentfolder"
+    mv * ../
+    cd ..
+    rm -r "$torrentfolder"
+fi
 
-videofile=$(find . -type f -exec file -N -i -- {} + | sed -n 's!: video/[^:]*$!!p')
+videofile="$(find . -type f -exec file -N -i -- {} + | sed -n 's!: video/[^:]*$!!p')"
+echo "Videofile: "$videofile
 
 echo "Downloading subtitles..."
-subliminal download -l en $videofile
+sudo subliminal download -l en "$videofile"
 
 subs=$(find | grep ".srt")
 
-read -p "Cast to chromecast? This will disconnect nordvpn. (y/n)" -n 1 -r
+read -p "Cast to chromecast? This will disconnect nordvpn. (y/n): " -n 1 -r
+echo ""
 
 if [[ $REPLY =~ ^[Yy]$ ]]
 then
-    nordvpn d > /dev/null
+    echo "Disconnecting nordvpn..."
+    nordvpn d
     clear
     castnow "$videofile" --subtitles "$subs" 
 fi
